+++ projects/client/src/components/Register/index.js
@@ -1,0 +1,3 @@
+function Register(){}
+
+export default Register;
